Organize research scripts
